changed long to int

+++ app/src/main/java/se/jakob/knarkklocka/TimePickerFragment.java
@@ -8,6 +8,7 @@ import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
 import android.widget.Chronometer;
 import android.widget.TimePicker;
+
 import static android.text.format.DateUtils.HOUR_IN_MILLIS;
 import static android.text.format.DateUtils.MINUTE_IN_MILLIS;
 
@@ -26,8 +27,8 @@ public class TimePickerFragment extends DialogFragment implements TimePickerDial
     public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
         long hoursInMillis = hourOfDay * HOUR_IN_MILLIS;
         long minutesInMillis = minute * MINUTE_IN_MILLIS;
-        long duration = hoursInMillis + minutesInMillis;
-        PreferenceUtils.setCustomTimerLength(getActivity(), duration);
+        long duration = (hoursInMillis + minutesInMillis);
+        PreferenceUtils.setCustomTimerLength(getActivity(), (int) duration);
         Chronometer chronometer = getActivity().findViewById(R.id.time_display);
         if (chronometer != null) {
             chronometer.setBase(SystemClock.elapsedRealtime() + duration);
